consolidate metadata imports
